% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(ISCO, type = "EGP", selfEmployed = NULL, unemployed = NULL)
}
\arguments{
\item{ISCO}{Vector of four digit ISCO-88 codes as int or cha}

\item{type}{The requested conversion type (see details)}

\item{selfEmployed}{The code that is used for self-employed individuals (see details)}

\item{unemployed}{The code that is used for unemployed individuals (see details)}
}
\description{

}
\details{
Converts ISCO-88 to Common SES Codes

Types include EGP, ISEI, and SIOPS. We strictly use Appendix A from Ganzeboom & Treiman (1996) for conversions. This means no distinction is made between Routine non-manual employees higher grade or lower grade. Both are coded as 3. ISCO-88 does not contain codes for self-employed or unemployed individuals. Thus user input is required to specify the codes used for each (if any). Self-employed and unemployed codes will be translated to EGP codes of 6 and 12 respectively. They will be coded as NA for ISEI or SIOPS.
}
\examples{
convert(c(1000,9000), type = "EGP")
}
