\name{ISCO0888}
\docType{data}
\alias{ISCO0888}
\title{Hash table with ISCO08 codes as Key and correspondance ISCO88 codes as Values}

\usage{ISCO0888}

\description{ISCO-08 and ISCO-88 correspondance table.}

\format{A hash table with 436 Key:Value pairs}

\source{
ABS
}

\details{
Key: ISCO 08 values

Values: ISCO 88 value
}

\keyword{datasets}

\examples{
#See what ISCO-08 codes are avaliable
hash::keys(ISCO0888)
}
